+++ CHANGELOG.md
@@ -14,6 +14,7 @@ We use *breaking :warning:* to mark changes that are not backward compatible (re
 ### Changed
 
 - [#3136](https://github.com/thanos-io/thanos/pull/3136) Sidecar: Add metric `thanos_sidecar_reloader_config_apply_operations_total` and rename metric `thanos_sidecar_reloader_config_apply_errors_total` to `thanos_sidecar_reloader_config_apply_operations_failed_total`.
+- [#3148](https://github.com/thanos-io/thanos/pull/3148) Tracing: Instrument memcached client.
 
 ## [v0.15.0](https://github.com/thanos-io/thanos/releases) - 2020.09.07